# openapi.yaml
openapi: "3.0.0"
info:
    version: "1.0.0"
    title: "Daily Tasks API"
    description: Describes the API of Daily Tasks
paths:
...
    /api/with-auth/tasks/backlog/{backlog_type}/for/{day}:
        description: Get all tasks in backlog
        parameters:
            - name: backlog_type
              in: path
              required: true
              schema:
                  $ref: "#/components/schemas/BacklogType"
            - name: day
              in: path
              required: true
              schema:
                  type: string
                  format: date
        get:
            summary: Get a list of tasks in a specific backlog
            responses:
                "200":
                    description: A list of tasks
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Task"

        post:
            summary: Add a task in a backlog
            requestBody:
                description: JSON payload
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/NewTaskData"
            responses:
                "200":
                    description: Created task data
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AddNewTaskToBacklogResponse"
                "401":
                    description: Unathorized
...
components:
    schemas:
...
        AddNewTaskToBacklogResponse:
            type: object
            required:
                - taskId
            properties:
                taskId:
                    type: integer
        Task:
            type: object
            required:
                - id
                - title
                - status
                - createdAt
                - updatedAt
            properties:
                id:
                    type: integer
                title:
                    type: string
                    minLength: 1
                description:
                    type: string
                status:
                    type: string
                    enum:
                        - not_started
                        - in_progress
                        - completed
                        - archived
                priority:
                    type: string
                    enum:
                        - low
                        - medium
                        - high
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        NewTaskData:
            type: object
            required:
                - title
            properties:
                title:
                    type: string
                    minLength: 1
                description:
                    type: string
                priority:
                    type: string
                    enum:
                        - low
                        - medium
                        - high
        BacklogType:
            type: string
            enum:
                - DAY
                - WEEK
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
security:
    - BearerAuth: []